Refactoring of c8yclient and moving parts of the implementation to shared/
